pyproject.toml
[project]
name = "ai-chat-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Sai Naveen Adep"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.128.3,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "qdrant-client (>=1.16.2,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
